% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{draw_column}
\alias{draw_column}
\title{Draw a column}
\usage{
draw_column(
  xpos,
  ncubes = 10,
  fill = "gray80",
  darkenby = 0.05,
  z = 0,
  delete_x = NULL,
  delete_y = NULL,
  light = "left-top",
  ysize = 1/ncubes
)
}
\arguments{
\item{xpos}{x}

\item{ncubes}{x}

\item{fill}{x}

\item{darkenby}{x}

\item{z}{x}

\item{delete_x}{x}

\item{delete_y}{x}

\item{light}{x}

\item{ysize}{x}
}
\value{
x
}
\description{
Draw a column
}
\examples{
\dontrun{
}
}
