% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{draw_row}
\alias{draw_row}
\title{Draw a row}
\usage{
draw_row(
  ypos,
  ncubes = 10,
  fill = "gray80",
  darkenby = 0.05,
  z = 0,
  delete = NULL,
  light = "left-top",
  ysize = 1/ncubes
)
}
\arguments{
\item{ypos}{x}

\item{ncubes}{x}

\item{fill}{x}

\item{darkenby}{x}

\item{z}{x}

\item{delete}{x}

\item{light}{x}

\item{ysize}{x}
}
\value{
x
}
\description{
Draw a row
}
\examples{
\dontrun{
}
}
