% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{draw_text}
\alias{draw_text}
\title{Draw text for labels}
\usage{
draw_text(
  coords,
  label = NULL,
  fill = NULL,
  fill2 = NULL,
  fill3 = NULL,
  light = "top-left",
  darkenby = 0.1,
  ysize = 1/25,
  xo = 0.5,
  yo = ysize,
  orient = "col",
  verbose = FALSE,
  fontsize = 9,
  just = "center",
  ...
)
}
\arguments{
\item{coords}{Dataframe of x,y,z coordinates}

\item{label}{Label vector}

\item{fill}{Fill color}

\item{fill2}{Fill color}

\item{fill3}{Fill color}

\item{light}{Light source char}

\item{darkenby}{Darken}

\item{ysize}{x}

\item{xo}{x}

\item{yo}{x}

\item{orient}{x}

\item{verbose}{x}

\item{fontsize}{x}

\item{just}{x}

\item{...}{x}
}
\value{
textgrob
}
\description{
Draw text for labels
}
\examples{
\dontrun{
}
}
