% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{draw_z}
\alias{draw_z}
\title{Draw on z}
\usage{
draw_z(zpos, ncubes = 10, fill = "gray80", darkenby = 0.05)
}
\arguments{
\item{zpos}{x}

\item{ncubes}{x}

\item{fill}{x}

\item{darkenby}{x}
}
\value{
x
}
\description{
Draw on z
}
\examples{
\dontrun{
}
}
