% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{drawcubes}
\alias{drawcubes}
\title{Draw a batch of prepared isocubes}
\usage{
drawcubes(obj, ...)
}
\arguments{
\item{obj}{A df with all the coordinate and fill information in it}

\item{...}{Arguments passed on to `isocubesGrob()`}
}
\value{
An isotib graphic
}
\description{
Take an object produced by e.g. `make_wall()` and draw it
}
\examples{
\dontrun{
}
}
