% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{get_labels}
\alias{get_labels}
\title{Get vector of labels}
\usage{
get_labels(labels, coords)
}
\arguments{
\item{labels}{x}

\item{coords}{x}
}
\value{
x
}
\description{
Get vector of labels
}
\examples{
\dontrun{
}
}
