% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocubes.R
\name{isocubesGrob}
\alias{isocubesGrob}
\title{Create a grob of isocubes}
\usage{
isocubesGrob(
  ...,
  coords,
  fill = "grey90",
  fill2 = NULL,
  fill3 = NULL,
  light = "top-left",
  darkenby = 0.3,
  ysize = 1/20,
  xo = 0.5,
  yo = ysize,
  verbose = FALSE
)
}
\arguments{
\item{...}{other values passed to \code{gpar()} to set the graphical
parameters e.g. \code{lwd} and \code{col} for the linewidth and colour
of the outline stroke for each cube face.}

\item{coords}{data.frame of x,y,z coordinates for the cubes (integer coordinates)}

\item{fill}{fill colour for the main face of cube as specified by the \code{light}
argument.  By default, this will be the colour of the top face of the cube,
as the default \code{light = 'top-left'}.}

\item{fill2, fill3}{fill colours for secondary and tertiary faces of
cube - as specified in the \code{light} argument.
If set to NULL (the default) then cube faces will be darkened versions
of the main \code{fill} color.}

\item{light}{direction of light.  This is a two-word argument nominating the
main light direction and secondary light direction. The default value of
'top-left' indicates the light is brightest from the top, and then
from the left side of the cube - with the right side of the cube being darkest.
This argument is only used if \code{fill2} and \code{fill3} are unspecified.
Possible values are: top-left, top-right, left-top, left-right, right-top,
right-left.}

\item{ysize}{the size of a cube as a fraction of the window height. Default: 0.05,
which means that, at most, 20 cubes could stack from top to bottom and be
visible.}

\item{xo, yo}{the origin of the isometric coordinate system in 'snpc' coordinates.
These values should be given as vanilla floating point values.
Be default the origin is the middle bottom of the graphics device
i.e. \code{(xo, yo) = (0.5, 0)}}

\item{verbose}{Be verbose? Default FALSE.}
}
\description{
Create a grob of isocubes
}
