% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{make_textvec}
\alias{make_textvec}
\title{Make a vector of text labels}
\usage{
make_textvec(labels, width, height, z = 0, color = "gray50")
}
\arguments{
\item{labels}{Character vector of labels}

\item{width}{Width of the wall}

\item{height}{Height of the wall}

\item{z}{Z-position}

\item{color}{Color}
}
\value{
A vector of labels
}
\description{
Make a vector of text labels
}
\examples{
\dontrun{
}
}
