% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{my_plot_crop}
\alias{my_plot_crop}
\title{Trim all border whitespace in a figure already saved to a file}
\usage{
my_plot_crop(x, quiet = TRUE)
}
\arguments{
\item{x}{Filename}

\item{quiet}{Shut up}
}
\value{
Converts the figure in place
}
\description{
Trim all border whitespace in a figure already saved to a file
}
\examples{
\dontrun{
}
}
