% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{paint_x}
\alias{paint_x}
\title{Paint x vec}
\usage{
paint_x(object, range, col)
}
\arguments{
\item{object}{x}

\item{range}{x}

\item{col}{x}
}
\value{
x
}
\description{
Paint x vec
}
\examples{
\dontrun{
}
}
