% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{make_brick}
\alias{make_brick}
\title{ISO coords}
\usage{
make_brick(width, height, depth, fill = "gray90")
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{depth}{block depth}

\item{fill}{Fill color}
}
\value{
Vector of brick coordinates
}
\description{
ISO coords
}
