% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fns.R
\name{make_wall}
\alias{make_wall}
\title{Make the grid points for an inso 'wall'}
\usage{
make_wall(width, height, z = 0, fill = "gray90")
}
\arguments{
\item{width}{n columns in the wall}

\item{height}{n rows in the wall}

\item{z}{z (3rd dimension) position of the wall}

\item{fill}{Color of the wall}
}
\value{
A dataframe of coordinates and color fill
}
\description{
Make the grid points for an inso 'wall'
}
