% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_fns.R
\name{paint_xy}
\alias{paint_xy}
\title{Paint xy}
\usage{
paint_xy(object, xrange, yrange, col)
}
\arguments{
\item{object}{x}

\item{xrange}{x}

\item{yrange}{x}

\item{col}{x}
}
\value{
x
}
\description{
Paint xy
}
